feat: check docs can be generated (#289)
ensure docs can be generated

+++ tools/DocGen/Program.fs
@@ -287,13 +287,20 @@ let writeExtension extensionDir (extension: Extension) =
 
 [<EntryPoint>]
 let main args =
-    if args.Length <> 2 then failwith "Usage: DocGen <xml-doc-file> <output-dir>"
-    let doc = load args[0]
-    let outputDir = args[1]
+    let outputDir, write =
+        match args with
+        | [| outputDir |] -> outputDir, false
+        | [| outputDir; "--write" |] -> outputDir, true
+        | _ -> failwith "Usage: DocGen <output-dir> [<--write>]"
+
+    let assemblyPath = Assembly.GetExecutingAssembly().Location
+    let docPath = Path.GetDirectoryName(assemblyPath) |> nonNull
+    let docFile = Path.Combine(docPath, "Terrabuild.Extensions.xml")
+    let doc = load docFile
     if doc.Assembly.Name <> "Terrabuild.Extensions" then failwith "Expecting documentation for Terrabuild.Extensions"
 
     let members = doc.Members |> Option.ofObj |> Option.defaultValue Array.empty
-    let assemblyFile = Path.ChangeExtension(args[0], "dll") |> nonNull
+    let assemblyFile = Path.ChangeExtension(docFile, "dll") |> nonNull
     let assembly = System.Reflection.Assembly.LoadFrom assemblyFile
     let extensions = buildExtensions assembly members
 
@@ -301,45 +308,46 @@ let main args =
     printfn "Generating docs"
     for (KeyValue(_, extension)) in extensions do
         let extensionDir = Path.Combine(outputDir, extension.Name)
-        if Directory.Exists extensionDir |> not then Directory.CreateDirectory extensionDir |> ignore
+        if write && Directory.Exists extensionDir |> not then Directory.CreateDirectory extensionDir |> ignore
 
         printfn $"  {extension.Name}"
-        writeExtension extensionDir extension
+        if write then writeExtension extensionDir extension
 
         // generate extension commands
         for cmd in extension.Commands do
             let batchCmd = extension.Commands |> List.tryFind (fun x -> x.Name = $"__{cmd.Name}__")
-            writeCommand extensionDir cmd batchCmd extension
+            if write then writeCommand extensionDir cmd batchCmd extension
 
     // cleanup
-    printfn "Cleaning output"
-    let genExtensions = extensions.Keys |> Set.ofSeq
-    let folders =
-        Directory.EnumerateDirectories(outputDir)
-        |> Seq.map (nonNull << Path.GetFileName)
-        |> Set.ofSeq
-    let removeFolders = folders - genExtensions
-    for folder in removeFolders do
-        let folder = Path.Combine(outputDir, folder)
-        printfn $"  Removing {folder}"
-        Directory.Delete(folder, true)
-
-    for (KeyValue(_, extension)) in extensions do
-        let extensionDir = Path.Combine(outputDir, extension.Name)
-        let genCommands =
-            extension.Commands
-            |> List.map (fun cmd -> $"{cmd.Name}.md")
-            |> Set.ofSeq
-            |> Set.add "_index.md"
-        let commands =
-            Directory.EnumerateFiles(extensionDir)
+    if write then
+        printfn "Cleaning output"
+        let genExtensions = extensions.Keys |> Set.ofSeq
+        let folders =
+            Directory.EnumerateDirectories(outputDir)
             |> Seq.map (nonNull << Path.GetFileName)
             |> Set.ofSeq
-        let removeCommands = commands - genCommands
-        for command in removeCommands do
-            let file = Path.Combine(extensionDir, command)
-            printfn $"  Removing {file}"
-            File.Delete(file)
+        let removeFolders = folders - genExtensions
+        for folder in removeFolders do
+            let folder = Path.Combine(outputDir, folder)
+            printfn $"  Removing {folder}"
+            Directory.Delete(folder, true)
+
+        for (KeyValue(_, extension)) in extensions do
+            let extensionDir = Path.Combine(outputDir, extension.Name)
+            let genCommands =
+                extension.Commands
+                |> List.map (fun cmd -> $"{cmd.Name}.md")
+                |> Set.ofSeq
+                |> Set.add "_index.md"
+            let commands =
+                Directory.EnumerateFiles(extensionDir)
+                |> Seq.map (nonNull << Path.GetFileName)
+                |> Set.ofSeq
+            let removeCommands = commands - genCommands
+            for command in removeCommands do
+                let file = Path.Combine(extensionDir, command)
+                printfn $"  Removing {file}"
+                File.Delete(file)
 
     printfn "Done"
     0
